Move Ctrl-D related code from contrib/find to contrib/multicursor
