+++ src/vs/editor/contrib/find/common/findModel.ts
@@ -50,10 +50,6 @@ export const FIND_IDS = {
 	PreviousMatchFindAction: 'editor.action.previousMatchFindAction',
 	NextSelectionMatchFindAction: 'editor.action.nextSelectionMatchFindAction',
 	PreviousSelectionMatchFindAction: 'editor.action.previousSelectionMatchFindAction',
-	AddSelectionToNextFindMatchAction: 'editor.action.addSelectionToNextFindMatch',
-	AddSelectionToPreviousFindMatchAction: 'editor.action.addSelectionToPreviousFindMatch',
-	MoveSelectionToNextFindMatchAction: 'editor.action.moveSelectionToNextFindMatch',
-	MoveSelectionToPreviousFindMatchAction: 'editor.action.moveSelectionToPreviousFindMatch',
 	StartFindReplaceAction: 'editor.action.startFindReplaceAction',
 	CloseFindWidgetCommand: 'closeFindWidget',
 	ToggleCaseSensitiveCommand: 'toggleFindCaseSensitive',
