+++ src/vs/editor/contrib/find/test/common/findController.test.ts
@@ -11,21 +11,16 @@ import { EditOperation } from 'vs/editor/common/core/editOperation';
 import { Position } from 'vs/editor/common/core/position';
 import { Selection } from 'vs/editor/common/core/selection';
 import { Range } from 'vs/editor/common/core/range';
-import { EndOfLineSequence, ICommonCodeEditor, Handler } from 'vs/editor/common/editorCommon';
-import {
-	CommonFindController, FindStartFocusAction, IFindStartOptions,
-	NextMatchFindAction, StartFindAction, SelectHighlightsAction,
-	AddSelectionToNextFindMatchAction,
-	MultiCursorSelectionController
-} from 'vs/editor/contrib/find/common/findController';
-import { MockCodeEditor, withMockCodeEditor } from 'vs/editor/test/common/mocks/mockCodeEditor';
+import { ICommonCodeEditor } from 'vs/editor/common/editorCommon';
+import { CommonFindController, FindStartFocusAction, IFindStartOptions, NextMatchFindAction, StartFindAction } from 'vs/editor/contrib/find/common/findController';
+import { withMockCodeEditor } from 'vs/editor/test/common/mocks/mockCodeEditor';
 import { HistoryNavigator } from 'vs/base/common/history';
 import { IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
 import { IStorageService } from 'vs/platform/storage/common/storage';
 import { ServiceCollection } from 'vs/platform/instantiation/common/serviceCollection';
 import { Delayer } from 'vs/base/common/async';
 
-class TestFindController extends CommonFindController {
+export class TestFindController extends CommonFindController {
 
 	public hasFocus: boolean;
 	public delayUpdateHistory: boolean = false;
@@ -189,62 +184,6 @@ suite('FindController', () => {
 		});
 	});
 
-	test('issue #8817: Cursor position changes when you cancel multicursor', () => {
-		withMockCodeEditor([
-			'var x = (3 * 5)',
-			'var y = (3 * 5)',
-			'var z = (3 * 5)',
-		], { serviceCollection: serviceCollection }, (editor, cursor) => {
-
-			let findController = editor.registerAndInstantiateContribution<TestFindController>(TestFindController);
-			let multiCursorSelectController = editor.registerAndInstantiateContribution<MultiCursorSelectionController>(MultiCursorSelectionController);
-			let selectHighlightsAction = new SelectHighlightsAction();
-
-			editor.setSelection(new Selection(2, 9, 2, 16));
-
-			selectHighlightsAction.run(null, editor);
-			assert.deepEqual(editor.getSelections().map(fromRange), [
-				[2, 9, 2, 16],
-				[1, 9, 1, 16],
-				[3, 9, 3, 16],
-			]);
-
-			editor.trigger('test', 'removeSecondaryCursors', null);
-
-			assert.deepEqual(fromRange(editor.getSelection()), [2, 9, 2, 16]);
-
-			multiCursorSelectController.dispose();
-			findController.dispose();
-		});
-	});
-
-	test('issue #5400: "Select All Occurrences of Find Match" does not select all if find uses regex', () => {
-		withMockCodeEditor([
-			'something',
-			'someething',
-			'someeething',
-			'nothing'
-		], { serviceCollection: serviceCollection }, (editor, cursor) => {
-
-			let findController = editor.registerAndInstantiateContribution<TestFindController>(TestFindController);
-			let selectHighlightsAction = new SelectHighlightsAction();
-
-			editor.setSelection(new Selection(1, 1, 1, 1));
-			findController.getState().change({ searchString: 'some+thing', isRegex: true, isRevealed: true }, false);
-
-			selectHighlightsAction.run(null, editor);
-			assert.deepEqual(editor.getSelections().map(fromRange), [
-				[1, 1, 1, 10],
-				[2, 1, 2, 11],
-				[3, 1, 3, 12],
-			]);
-
-			assert.equal(findController.getState().searchString, 'some+thing');
-
-			findController.dispose();
-		});
-	});
-
 	test('issue #9043: Clear search scope when find widget is hidden', () => {
 		withMockCodeEditor([
 			'var x = (3 * 5)',
@@ -383,125 +322,6 @@ suite('FindController', () => {
 		});
 	});
 
-	test('AddSelectionToNextFindMatchAction can work with multiline', () => {
-		withMockCodeEditor([
-			'',
-			'qwe',
-			'rty',
-			'',
-			'qwe',
-			'',
-			'rty',
-			'qwe',
-			'rty'
-		], { serviceCollection: serviceCollection }, (editor, cursor) => {
-
-			let findController = editor.registerAndInstantiateContribution<TestFindController>(TestFindController);
-			let multiCursorSelectController = editor.registerAndInstantiateContribution<MultiCursorSelectionController>(MultiCursorSelectionController);
-			let addSelectionToNextFindMatch = new AddSelectionToNextFindMatchAction();
-
-			editor.setSelection(new Selection(2, 1, 3, 4));
-
-			addSelectionToNextFindMatch.run(null, editor);
-			assert.deepEqual(editor.getSelections().map(fromRange), [
-				[2, 1, 3, 4],
-				[8, 1, 9, 4]
-			]);
-
-			editor.trigger('test', 'removeSecondaryCursors', null);
-
-			assert.deepEqual(fromRange(editor.getSelection()), [2, 1, 3, 4]);
-
-			multiCursorSelectController.dispose();
-			findController.dispose();
-		});
-	});
-
-	test('issue #6661: AddSelectionToNextFindMatchAction can work with touching ranges', () => {
-		withMockCodeEditor([
-			'abcabc',
-			'abc',
-			'abcabc',
-		], { serviceCollection: serviceCollection }, (editor, cursor) => {
-
-			let findController = editor.registerAndInstantiateContribution<TestFindController>(TestFindController);
-			let multiCursorSelectController = editor.registerAndInstantiateContribution<MultiCursorSelectionController>(MultiCursorSelectionController);
-			let addSelectionToNextFindMatch = new AddSelectionToNextFindMatchAction();
-
-			editor.setSelection(new Selection(1, 1, 1, 4));
-
-			addSelectionToNextFindMatch.run(null, editor);
-			assert.deepEqual(editor.getSelections().map(fromRange), [
-				[1, 1, 1, 4],
-				[1, 4, 1, 7]
-			]);
-
-			addSelectionToNextFindMatch.run(null, editor);
-			addSelectionToNextFindMatch.run(null, editor);
-			addSelectionToNextFindMatch.run(null, editor);
-			assert.deepEqual(editor.getSelections().map(fromRange), [
-				[1, 1, 1, 4],
-				[1, 4, 1, 7],
-				[2, 1, 2, 4],
-				[3, 1, 3, 4],
-				[3, 4, 3, 7]
-			]);
-
-			editor.trigger('test', Handler.Type, { text: 'z' });
-			assert.deepEqual(editor.getSelections().map(fromRange), [
-				[1, 2, 1, 2],
-				[1, 3, 1, 3],
-				[2, 2, 2, 2],
-				[3, 2, 3, 2],
-				[3, 3, 3, 3]
-			]);
-			assert.equal(editor.getValue(), [
-				'zz',
-				'z',
-				'zz',
-			].join('\n'));
-
-			multiCursorSelectController.dispose();
-			findController.dispose();
-		});
-	});
-
-	test('issue #23541: Multiline Ctrl+D does not work in CRLF files', () => {
-		withMockCodeEditor([
-			'',
-			'qwe',
-			'rty',
-			'',
-			'qwe',
-			'',
-			'rty',
-			'qwe',
-			'rty'
-		], { serviceCollection: serviceCollection }, (editor, cursor) => {
-
-			editor.getModel().setEOL(EndOfLineSequence.CRLF);
-
-			let findController = editor.registerAndInstantiateContribution<TestFindController>(TestFindController);
-			let multiCursorSelectController = editor.registerAndInstantiateContribution<MultiCursorSelectionController>(MultiCursorSelectionController);
-			let addSelectionToNextFindMatch = new AddSelectionToNextFindMatchAction();
-
-			editor.setSelection(new Selection(2, 1, 3, 4));
-
-			addSelectionToNextFindMatch.run(null, editor);
-			assert.deepEqual(editor.getSelections().map(fromRange), [
-				[2, 1, 3, 4],
-				[8, 1, 9, 4]
-			]);
-
-			editor.trigger('test', 'removeSecondaryCursors', null);
-
-			assert.deepEqual(fromRange(editor.getSelection()), [2, 1, 3, 4]);
-
-			multiCursorSelectController.dispose();
-			findController.dispose();
-		});
-	});
-
 	test('issue #18111: Regex replace with single space replaces with no space', () => {
 		withMockCodeEditor([
 			'HRESULT OnAmbientPropertyChange(DISPID   dispid);'
@@ -564,239 +384,6 @@ suite('FindController', () => {
 		}
 		return result;
 	}
-
-	function testAddSelectionToNextFindMatchAction(text: string[], callback: (editor: MockCodeEditor, action: AddSelectionToNextFindMatchAction, findController: TestFindController) => void): void {
-		withMockCodeEditor(text, { serviceCollection: serviceCollection }, (editor, cursor) => {
-
-			let findController = editor.registerAndInstantiateContribution<TestFindController>(TestFindController);
-			let multiCursorSelectController = editor.registerAndInstantiateContribution<MultiCursorSelectionController>(MultiCursorSelectionController);
-
-			let action = new AddSelectionToNextFindMatchAction();
-
-			callback(editor, action, findController);
-
-			multiCursorSelectController.dispose();
-			findController.dispose();
-		});
-	}
-
-	test('AddSelectionToNextFindMatchAction starting with single collapsed selection', () => {
-		const text = [
-			'abc pizza',
-			'abc house',
-			'abc bar'
-		];
-		testAddSelectionToNextFindMatchAction(text, (editor, action, findController) => {
-			editor.setSelections([
-				new Selection(1, 2, 1, 2),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 4),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 4),
-				new Selection(2, 1, 2, 4),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 4),
-				new Selection(2, 1, 2, 4),
-				new Selection(3, 1, 3, 4),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 4),
-				new Selection(2, 1, 2, 4),
-				new Selection(3, 1, 3, 4),
-			]);
-		});
-	});
-
-	test('AddSelectionToNextFindMatchAction starting with two selections, one being collapsed 1)', () => {
-		const text = [
-			'abc pizza',
-			'abc house',
-			'abc bar'
-		];
-		testAddSelectionToNextFindMatchAction(text, (editor, action, findController) => {
-			editor.setSelections([
-				new Selection(1, 1, 1, 4),
-				new Selection(2, 2, 2, 2),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 4),
-				new Selection(2, 1, 2, 4),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 4),
-				new Selection(2, 1, 2, 4),
-				new Selection(3, 1, 3, 4),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 4),
-				new Selection(2, 1, 2, 4),
-				new Selection(3, 1, 3, 4),
-			]);
-		});
-	});
-
-	test('AddSelectionToNextFindMatchAction starting with two selections, one being collapsed 2)', () => {
-		const text = [
-			'abc pizza',
-			'abc house',
-			'abc bar'
-		];
-		testAddSelectionToNextFindMatchAction(text, (editor, action, findController) => {
-			editor.setSelections([
-				new Selection(1, 2, 1, 2),
-				new Selection(2, 1, 2, 4),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 4),
-				new Selection(2, 1, 2, 4),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 4),
-				new Selection(2, 1, 2, 4),
-				new Selection(3, 1, 3, 4),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 4),
-				new Selection(2, 1, 2, 4),
-				new Selection(3, 1, 3, 4),
-			]);
-		});
-	});
-
-	test('AddSelectionToNextFindMatchAction starting with all collapsed selections', () => {
-		const text = [
-			'abc pizza',
-			'abc house',
-			'abc bar'
-		];
-		testAddSelectionToNextFindMatchAction(text, (editor, action, findController) => {
-			editor.setSelections([
-				new Selection(1, 2, 1, 2),
-				new Selection(2, 2, 2, 2),
-				new Selection(3, 1, 3, 1),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 4),
-				new Selection(2, 1, 2, 4),
-				new Selection(3, 1, 3, 4),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 4),
-				new Selection(2, 1, 2, 4),
-				new Selection(3, 1, 3, 4),
-			]);
-		});
-	});
-
-	test('AddSelectionToNextFindMatchAction starting with all collapsed selections on different words', () => {
-		const text = [
-			'abc pizza',
-			'abc house',
-			'abc bar'
-		];
-		testAddSelectionToNextFindMatchAction(text, (editor, action, findController) => {
-			editor.setSelections([
-				new Selection(1, 6, 1, 6),
-				new Selection(2, 6, 2, 6),
-				new Selection(3, 6, 3, 6),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 5, 1, 10),
-				new Selection(2, 5, 2, 10),
-				new Selection(3, 5, 3, 8),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 5, 1, 10),
-				new Selection(2, 5, 2, 10),
-				new Selection(3, 5, 3, 8),
-			]);
-		});
-	});
-
-	test('issue #20651: AddSelectionToNextFindMatchAction case insensitive', () => {
-		const text = [
-			'test',
-			'testte',
-			'Test',
-			'testte',
-			'test'
-		];
-		testAddSelectionToNextFindMatchAction(text, (editor, action, findController) => {
-			editor.setSelections([
-				new Selection(1, 1, 1, 5),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 5),
-				new Selection(2, 1, 2, 5),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 5),
-				new Selection(2, 1, 2, 5),
-				new Selection(3, 1, 3, 5),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 5),
-				new Selection(2, 1, 2, 5),
-				new Selection(3, 1, 3, 5),
-				new Selection(4, 1, 4, 5),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 5),
-				new Selection(2, 1, 2, 5),
-				new Selection(3, 1, 3, 5),
-				new Selection(4, 1, 4, 5),
-				new Selection(5, 1, 5, 5),
-			]);
-
-			action.run(null, editor);
-			assert.deepEqual(editor.getSelections(), [
-				new Selection(1, 1, 1, 5),
-				new Selection(2, 1, 2, 5),
-				new Selection(3, 1, 3, 5),
-				new Selection(4, 1, 4, 5),
-				new Selection(5, 1, 5, 5),
-			]);
-		});
-	});
 });
 
 suite('FindController query options persistence', () => {
